+++ src/main/java/com/lukaskucera/numberneighbors/controller/PlayerController.java
@@ -3,12 +3,14 @@ package com.lukaskucera.numberneighbors.controller;
 import com.lukaskucera.numberneighbors.entity.Player;
 import com.lukaskucera.numberneighbors.request.NewPlayerRequest;
 import com.lukaskucera.numberneighbors.response.NewPlayerResponse;
+import com.lukaskucera.numberneighbors.service.GameServiceImpl;
 import com.lukaskucera.numberneighbors.service.JwtService;
 import com.lukaskucera.numberneighbors.service.PlayerServiceImpl;
 import java.util.Set;
 import org.jetbrains.annotations.NotNull;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.oauth2.server.resource.authentication.JwtAuthenticationToken;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -20,16 +22,21 @@ import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 public class PlayerController {
+  private final GameServiceImpl gameService;
   private final PlayerServiceImpl playerService;
   private final JwtService jwtService;
 
-  public PlayerController(PlayerServiceImpl playerService, JwtService jwtService) {
+  public PlayerController(
+      GameServiceImpl gameService, PlayerServiceImpl playerService, JwtService jwtService) {
+    this.gameService = gameService;
     this.playerService = playerService;
     this.jwtService = jwtService;
   }
 
   @GetMapping(value = "/players")
-  public ResponseEntity<Set<Player>> players(@RequestParam(name = "game") Long gameId) {
+  public ResponseEntity<Set<Player>> players(
+      @RequestParam(name = "game") Long gameId, @NotNull JwtAuthenticationToken jwtToken) {
+    gameService.checkGameAccess(gameId, jwtToken);
     return ResponseEntity.ok(playerService.getPlayersByGameId(gameId));
   }
 
@@ -44,13 +51,16 @@ public class PlayerController {
   }
 
   @GetMapping(value = "/players/{id}")
-  public ResponseEntity<Player> player(@PathVariable Long id) {
+  public ResponseEntity<Player> player(
+      @PathVariable Long id, @NotNull JwtAuthenticationToken jwtToken) {
+    playerService.checkPlayerAccess(id, jwtToken);
     return ResponseEntity.ok(playerService.getPlayerById(id));
   }
 
   @DeleteMapping(value = "/players/{id}")
   @ResponseStatus(value = HttpStatus.NO_CONTENT)
-  public void deletePlayer(@PathVariable Long id) {
+  public void deletePlayer(@PathVariable Long id, @NotNull JwtAuthenticationToken jwtToken) {
+    playerService.checkPlayerAccess(id, jwtToken);
     playerService.deletePlayerById(id);
   }
 }
